+++ src/Contracts/Form/FactoryField.php
@@ -11,7 +11,7 @@ interface FactoryField extends FactoryResolver, ParamsBag
      *
      * @return string
      */
-    public function __toString();
+    public function __toString(): string;
 
     /**
      * Définition d'un message d'erreur associé au champ.
@@ -47,78 +47,78 @@ interface FactoryField extends FactoryResolver, ParamsBag
      *
      * @return mixed
      */
-    public function getAddonOption($name, $key = null, $default = null);
+    public function getAddonOption(string $name, ?string $key = null, $default = null);
 
     /**
      * Récupération de l'instance du contrôleur de champ.
      *
      * @return FieldController
      */
-    public function getController();
+    public function getController(): FieldController;
 
     /**
      * Récupération d'un ou de la liste des attributs de configuration complémentaires.
      *
-     * @param null|string $key Clé d'indexe de l'attribut. Syntaxe à point permise. Laisser à null (défaut) pour
+     * @param string|null $key Clé d'indexe de l'attribut. Syntaxe à point permise. Laisser à null (défaut) pour
      *                         récupérer la liste complète.
      * @param mixed $default Valeur de retour par défaut.
      *
      * @return mixed
      */
-    public function getExtras($key = null, $default = null);
+    public function getExtras(?string $key = null, $default = null);
 
     /**
      * Récupération du groupe d'appartenance.
      *
      * @return FactoryGroup
      */
-    public function getGroup();
+    public function getGroup(): ?FactoryGroup;
 
     /**
      * Récupération de l'indice de qualification de la variable de requête.
      *
      * @return string
      */
-    public function getName();
+    public function getName(): string;
 
     /**
      * Récupération de l'ordre d'affichage.
      *
      * @return int
      */
-    public function getPosition();
+    public function getPosition(): int;
 
     /**
      * Récupération d'attribut de champ requis.
      *
-     * @param null|string $key Clé d'indexe d'attributs. Syntaxe à point permise.
+     * @param string|null $key Clé d'indexe d'attributs. Syntaxe à point permise.
      *                         Retour la liste complète si null (défaut).
      * @param mixed $default Valeur de retour par défaut.
      *
      * @return mixed
      */
-    public function getRequired($key = null, $default = null);
+    public function getRequired(?string $key = null, $default = null);
 
     /**
      * Récupération de l'identifiant de qualification.
      *
      * @return string
      */
-    public function getSlug();
+    public function getSlug(): string;
 
     /**
      * Récupération de l'intitulé de qualification.
      *
      * @return string
      */
-    public function getTitle();
+    public function getTitle(): string;
 
     /**
      * Récupération du type.
      *
      * @return string
      */
-    public function getType();
+    public function getType(): string;
 
     /**
      * Récupération de la valeur.
@@ -127,76 +127,76 @@ interface FactoryField extends FactoryResolver, ParamsBag
      *
      * @return mixed
      */
-    public function getValue($raw = true);
+    public function getValue(bool $raw = true);
 
     /**
      * Récupération de la liste des valeurs.
      *
      * @param bool $raw Activation de la valeur de retour au format brut.
-     * @param null|string $glue Caractère d'assemblage de la valeur.
+     * @param string|null $glue Caractère(s) d'assemblage de la valeur.
      *
      * @return string|array
      */
-    public function getValues($raw = true, $glue = ', ');
+    public function getValues(bool $raw = true, ?string $glue = ', ');
 
     /**
      * Vérification d'existance d'une étiquette.
      *
      * @return boolean
      */
-    public function hasLabel();
+    public function hasLabel(): bool;
 
     /**
      * Vérification d'existance d'encapsuleur HTML.
      *
      * @return boolean
      */
-    public function hasWrapper();
+    public function hasWrapper(): bool;
 
     /**
      * Vérification si le champ retourne des erreurs de traitement.
      *
      * @return boolean
      */
-    public function onError();
+    public function onError(): bool;
 
     /**
      * Traitement récursif des tests de validation.
      *
-     * @param mixed $validations Test de validation à traiter.
+     * @param string|array $validations Test de validation à traiter.
      * @param array $results Liste des tests de validations traités.
      *
      * @return array
      */
-    public function parseValidations($validations, $results = []);
+    public function parseValidations($validations, array $results = []): array;
 
     /**
      * Initialisation (préparation) du champ.
      *
-     * @return void
+     * @return static
      */
-    public function prepare();
+    public function prepare(): FactoryField;
 
     /**
      * Réinitionalisation  de la valeur.
      *
-     * @return self
+     * @return static
      */
-    public function resetValue();
+    public function resetValue(): FactoryField;
 
     /**
      * Rendu de l'affichage.
      *
      * @return string
      */
-    public function render();
+    public function render(): string;
 
     /**
      * Préparation du rendu de l'affichage.
      *
-     * @return void
+     * @return static
      */
-    public function renderPrepare();
+    public function renderPrepare(): FactoryField;
 
     /**
      * Définition d'une attributs de configuration complémentaire.
@@ -204,9 +204,9 @@ interface FactoryField extends FactoryResolver, ParamsBag
      * @param string $key Clé d'indexe de l'attribut.
      * @param mixed $value Valeur à définir.
      *
-     * @return array
+     * @return static
      */
-    public function setExtra($key, $value);
+    public function setExtra(string $key, $value): FactoryField;
 
     /**
      * Définition du statut de l'indicateur de champ en erreur.
@@ -224,23 +224,30 @@ interface FactoryField extends FactoryResolver, ParamsBag
      *
      * @return $this
      */
-    public function setPosition($position = 0);
+    public function setPosition(int $position = 0): FactoryField;
 
     /**
-     * Définition de la valeur d'un champ.
+     * Définition de la valeur du champ basée sur la donnée en correspondance stockée en session.
+     *
+     * @return static
+     */
+    public function setSessionValue(): FactoryField;
+
+    /**
+     * Définition de la valeur du champ.
      *
      * @param mixed $value Valeur à définir.
      *
-     * @return $this
+     * @return static
      */
-    public function setValue($value);
+    public function setValue($value): FactoryField;
 
     /**
-     * Vérification d'une propriété ou récupération de la liste des proriétés de support .
+     * Vérification d'une propriété|récupération de la liste des proriétés de support .
      *
-     * @param null|string $support Propriété du support à vérifier.
+     * @param string|null $support Propriété du support à vérifier.
      *
      * @return array|boolean
      */
-    public function supports($support = null);
+    public function supports(?string $support = null);
 }