+++ src/Contracts/Form/FactoryResolver.php
@@ -7,7 +7,7 @@ use tiFy\Contracts\View\Engine as ViewEngine;
 interface FactoryResolver
 {
     /**
-     * Récupération de l'instance du contrôleur d'un addon associé au formulaire.
+     * Récupération de l'instance d'un addon
      *
      * @param string $name Nom de qualification de l'addon.
      *
@@ -16,21 +16,21 @@ interface FactoryResolver
     public function addon($name);
 
     /**
-     * Récupération de l'instance du contrôleur des addons associés au formulaire.
+     * Récupération de l'instance du gestionnaire d'addons.
      *
      * @return FactoryAddons|AddonFactory[]
      */
     public function addons();
 
     /**
-     * Récupération de l'instance du contrôleur des boutons associés au formulaire.
+     * Récupération de l'instance du gestionnaire de boutons.
      *
      * @return FactoryButtons|ButtonController[]
      */
     public function buttons();
 
     /**
-     * Récupération de l'instance du contrôleur des événements associés au formulaire ou déclenchement d'un événement.
+     * Récupération de l'instance du gestionnaire d'événements|déclenchement d'un événement.
      *
      * @param string $name Nom de qualification de l'événement.
      * @param array $args Liste des arguments complémentaires de déclenchement
@@ -40,7 +40,7 @@ interface FactoryResolver
     public function events($name = null, array $args = []);
 
     /**
-     * Récupération de l'instance du contrôleur d'un champ associé au formulaire.
+     * Récupération de l'instance d'un champ.
      *
      * @param null|string $slug Identifiant de qualification du champ.
      *
@@ -49,28 +49,42 @@ interface FactoryResolver
     public function field($slug = null);
 
     /**
-     * Récupération de l'instance du contrôleur des champs associés au formulaire.
+     * Récupération de l'instance du gestionnaire de champs.
      *
      * @return FactoryFields|FactoryField[]
      */
     public function fields();
 
     /**
-     * Récupération de l'instance du contrôleur de formulaire associé.
+     * Récupération de l'instance du gestionnaire de formulaire.
      *
      * @return FormFactory
      */
     public function form();
 
     /**
-     * Récupération de l'instance du contrôleur des messages de notification associés au formulaire.
+     * Récupération de l'instance du gestionnaire de groupes.
+     *
+     * @return FactoryGroups|FactoryGroup[]
+     */
+    public function groups();
+
+    /**
+     * Récupération de l'instance du gestionnaire de traitement de requête de soumission.
+     *
+     * @return FactoryHandle
+     */
+    public function handle();
+
+    /**
+     * Récupération de l'instance du gestionnaire des messages de notification.
      *
      * @return FactoryNotices
      */
     public function notices();
 
     /**
-     * Récupération d'une option ou de la liste complète des options du formulaire.
+     * Récupération d'une option|Liste complète des options.
      *
      * @param null|string $key Clé d'indice de l'option.
      * @param mixed $default Valeur de retour par défaut.
@@ -80,21 +94,14 @@ interface FactoryResolver
     public function option($key = null, $default = null);
 
     /**
-     * Récupération de l'instance du contrôleur des options associées au formulaire.
+     * Récupération de l'instance du gestionnaire d'options.
      *
      * @return FactoryOptions
      */
     public function options();
 
     /**
-     * Récupération de l'instance du contrôleur de traitement de la requête de soumission associée au formulaire.
-     *
-     * @return FactoryRequest
-     */
-    public function request();
-
-    /**
-     * Résolution de service fournis par le contructeur de formulaire.
+     * Résolution de service fourni.
      *
      * @param string $alias Nom de qualification du service.
      * @param array $args Liste des variables passées en argument au service.
@@ -104,21 +111,21 @@ interface FactoryResolver
     public function resolve($alias, $args = []);
 
     /**
-     * Récupération de l'instance du contrôleur de session associé au formulaire.
+     * Récupération de l'instance du gestionnaire de session.
      *
      * @return FactorySession
      */
     public function session();
 
     /**
-     * Récupération de l'instance du contrôleur de validation associé au formulaire.
+     * Récupération de l'instance du gestionnaire.
      *
      * @return FactoryValidation
      */
     public function validation();
 
     /**
-     * Récupération d'un instance du controleur de liste des gabarits d'affichage ou d'un gabarit d'affichage.
+     * Récupération d'une instance du gestionnaire de gabarits d'affichage|Contenu du gabarit d'affichage.
      * {@internal
      *  - cas 1 : Aucun argument n'est passé à la méthode, retourne l'instance du controleur de gabarit d'affichage.
      *  - cas 2 : Rétourne le gabarit d'affichage en passant les variables en argument.