+++ src/PostType/PostTypePostMeta.php
@@ -67,7 +67,7 @@ class PostTypePostMeta implements PostTypePostMetaContract
             $metas = [];
 
             foreach ((array)$res as $args) {
-                $metas[$args->meta_id] = Str::unserialize($args->meta_value);
+                $metas[$args->meta_id] = Str::unserialize($args->meta_value ?:'');
             }
 
             return $metas;