+++ src/Taxonomy/TaxonomyTermMeta.php
@@ -70,7 +70,7 @@ class TaxonomyTermMeta implements TaxonomyTermMetaContract
             $metas = [];
 
             foreach ((array)$res as $args) {
-                $metas[$args->meta_id] = Str::unserialize($args->meta_value);
+                $metas[$args->meta_id] = Str::unserialize($args->meta_value ?: '');
             }
 
             return $metas;